refactor: modularize PDF and certificate generation logic
